api-rest en proceso de montaje. Se muestran los municipios

+++ src/main/java/com/dalila/entity/Municipio.java
@@ -11,7 +11,7 @@ public class Municipio {
         this.nombre = nombre;
     }
 
-    public int getId() { return id; }
+    public long getId() { return id; }
     public void setId(int id) { this.id = id; }
 
     public String getNombre() { return nombre; }
